+++ packages/shared-ui/components/Search/index.ts
@@ -1,0 +1,1 @@
+export * from './Search'
