+++ packages/shared-ui/package.json
@@ -22,6 +22,7 @@
     "classnames": "^2.3.2",
     "downshift": "^6.1.9",
     "framer-motion": "^7.4.0",
+    "lodash.debounce": "^4.0.8",
     "nuka-carousel": "^5.3.0",
     "react-icons": "^4.4.0"
   },
@@ -33,6 +34,7 @@
     "@storybook/addon-a11y": "^7.5.2",
     "@storybook/addon-actions": "^7.5.2",
     "@storybook/react-vite": "^7.5.2",
+    "@types/lodash.debounce": "^4.0.7",
     "http-server": "^14.1.1",
     "react": "18.2.0",
     "react-dom": "18.2.0",
