+++ packages/shared-ui/tsconfig.json
@@ -12,5 +12,5 @@
     "node_modules",
     "dist",
     ".storybook"
-  ]
+  ],
 }
