+++ tsconfig.json
@@ -13,5 +13,9 @@
     "esModuleInterop": true,
     "module": "CommonJS",
     "moduleResolution": "node"
-  }
+  },
+  "include": [
+    "**/*.ts",
+    "**/*.tsx"
+  ],
 }
